Fix PCM audio streaming issues: frame alignment, noise, and skipping
- server: ensure PCM data chunks sent to clients are frame-aligned (4 bytes) to prevent decoding errors
- server: increase broadcast channel buffer size (100 -> 500) to mitigate audio skipping from packet drops
- server: change ffmpeg log level to 'error' to suppress non-critical HLS warnings
- client: fix incorrect sign extension in PCM sample reconstruction logic that caused noisy audio output

+++ player/http_player.go
@@ -147,7 +147,9 @@ func (vr *HTTPVolumeReader) Read(p []byte) (n int, err error) {
 		volume := vr.player.getEffectiveVolume()
 
 		for i := 0; i < n-1; i += 2 {
-			sample := int16(p[i]) | int16(p[i+1])<<8
+			// Correctly reconstruct signed 16-bit little-endian sample:
+			// Use uint16 for byte concatenation to avoid sign extension issues
+			sample := int16(uint16(p[i]) | uint16(p[i+1])<<8)
 			sample = int16(float64(sample) * volume)
 			p[i] = byte(sample)
 			p[i+1] = byte(sample >> 8)
